+++ CHANGELOG.md
@@ -6,9 +6,12 @@
 
 - `ModelArrows`
 - `ModelText`
+- `Transform.Inverse()`
+
 ### Changed
 
 ### Fixed
+
 - Deduplicate catalog names during code generation.
 
 ## 0.9.2
@@ -43,7 +46,6 @@
   - `Line.ExtendTo(Polygon)`
   - `Line.ExtendTo(Profile)`
 - Support for DXF from many basic elements.
-
 
 ### Changed